ScientificNameLabel on VC has been properly constrained to parent container so that it will scale when scrolled up. Also fixed access control all throughout the app for better OOP practices.
