+++ Terra/Terra/Controllers/SpeciesListViewController.swift
@@ -8,11 +8,11 @@
 
 import UIKit
 
-class SpeciesListViewController: UIViewController {
+final class SpeciesListViewController: UIViewController {
     
     var scrollView = UIScrollView()
     
-    lazy var terraTitleLabel: UILabel = {
+    private lazy var terraTitleLabel: UILabel = {
         let label = UILabel()
         label.text = "Terra"
         label.font = UIFont(name: "Roboto-Bold", size: 30)
@@ -21,7 +21,7 @@ class SpeciesListViewController: UIViewController {
         return label
     }()
     
-    lazy var subtitleLabel: UILabel = {
+    private lazy var subtitleLabel: UILabel = {
         let label = UILabel()
         label.text = "Protect the earth's biodiversity"
         label.font = UIFont(name: "Roboto-Light", size: 20)
@@ -30,7 +30,7 @@ class SpeciesListViewController: UIViewController {
         return label
     }()
     
-    lazy var criticalSpeciesLabel: UILabel = {
+    private lazy var criticalSpeciesLabel: UILabel = {
         let label = UILabel()
         label.text = "CRITICALLY ENDANGERED"
         label.font = UIFont(name: "Roboto-Medium", size: 19)
@@ -39,7 +39,7 @@ class SpeciesListViewController: UIViewController {
         return label
     }()
     
-    lazy var vulnerableSpeciesLabel: UILabel = {
+    private lazy var vulnerableSpeciesLabel: UILabel = {
         let label = UILabel()
         label.text = "VULNERABLE"
         label.font = UIFont(name: "Roboto-Medium", size: 19)
@@ -48,7 +48,7 @@ class SpeciesListViewController: UIViewController {
         return label
     }()
     
-    lazy var endangeredSpeciesLabel: UILabel = {
+    private lazy var endangeredSpeciesLabel: UILabel = {
         let label = UILabel()
         label.text = "ENDANGERED"
         label.font = UIFont(name: "Roboto-Medium", size: 21)
@@ -57,20 +57,20 @@ class SpeciesListViewController: UIViewController {
         return label
     }()
     
-    lazy var criticalSpeciesCV: UICollectionView = {
+    private lazy var criticalSpeciesCV: UICollectionView = {
         Utilities.makeCollectionView(superView: self.view)
     }()
     
-    lazy var endangeredSpeciesCV: UICollectionView = {
+    private lazy var endangeredSpeciesCV: UICollectionView = {
         Utilities.makeCollectionView(superView: self.view)
     }()
     
-    lazy var vulnerableSpeciesCV: UICollectionView = {
+    private lazy var vulnerableSpeciesCV: UICollectionView = {
         Utilities.makeCollectionView(superView: self.view)
     }()
     
     
-    var animalData = [Species]() {
+    private var animalData = [Species]() {
         didSet {
             filteredCriticalSpecies = filterSpecies(by: .critical)
             filteredEndangeredSpecies = filterSpecies(by: .endangered)
@@ -78,19 +78,19 @@ class SpeciesListViewController: UIViewController {
         }
     }
     
-    var filteredCriticalSpecies = [Species]() {
+    private var filteredCriticalSpecies = [Species]() {
         didSet {
             criticalSpeciesCV.reloadData()
         }
     }
     
-    var filteredEndangeredSpecies = [Species]() {
+    private var filteredEndangeredSpecies = [Species]() {
         didSet {
             endangeredSpeciesCV.reloadData()
         }
     }
     
-    var filteredVulnerableSpecies = [Species]() {
+    private var filteredVulnerableSpecies = [Species]() {
         didSet {
             vulnerableSpeciesCV.reloadData()
         }
@@ -120,6 +120,7 @@ class SpeciesListViewController: UIViewController {
         let collectionViews = [criticalSpeciesCV, endangeredSpeciesCV, vulnerableSpeciesCV]
         collectionViews.forEach { $0.dataSource = self }
         collectionViews.forEach { $0.delegate = self }
+        scrollView.delegate = self
     }
     
     
@@ -136,122 +137,11 @@ class SpeciesListViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = #colorLiteral(red: 0.1046695188, green: 0.09944508225, blue: 0.2029559612, alpha: 1)
+        addSubviews()
         setConstraints()
         loadSpeciesDataFromFirebase()
         setDatasourceAndDelegates()
-       
-        scrollView.delegate = self
     }
-}
-
-//MARK: -- Constraints
-extension SpeciesListViewController {
-    private func setConstraints(){
-        
-        let parentViewUIElements = [terraTitleLabel, subtitleLabel, scrollView]
-        parentViewUIElements.forEach { view.addSubview($0) }
-        parentViewUIElements.forEach { $0.translatesAutoresizingMaskIntoConstraints = false }
-        
-        let scrollViewUIElements = [criticalSpeciesLabel, criticalSpeciesCV, endangeredSpeciesLabel, endangeredSpeciesCV, vulnerableSpeciesLabel, vulnerableSpeciesCV]
-        scrollViewUIElements.forEach{ scrollView.addSubview($0) }
-        scrollViewUIElements.forEach{ $0.translatesAutoresizingMaskIntoConstraints = false }
-        
-        setTerraTitleLabelConstraints()
-        setSubtitleLabelConstraints()
-        
-        setCriticalSpeciesLabelConstraints()
-        setCriticalSpeciesCVConstraints()
-        
-        setEndangeredSpeciesLabelConstraints()
-        setEndangeredSpeciesCVConstraints()
-
-        setVulnerableSpeciesLabelConstraints()
-        setVulnerableSpeciesCVConstraints()
-    }
-    
-    private func setScrollViewConstraints(){
-         scrollView.backgroundColor = .clear
-         NSLayoutConstraint.activate([
-             scrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-             scrollView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-             scrollView.topAnchor.constraint(equalTo: subtitleLabel.bottomAnchor, constant: 30),
-             scrollView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-         ])
-     }
-     
-    
-    private func setTerraTitleLabelConstraints() {
-        NSLayoutConstraint.activate([
-            terraTitleLabel.topAnchor.constraint(equalTo: view.topAnchor, constant: 80),
-            terraTitleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 30),
-            terraTitleLabel.heightAnchor.constraint(equalToConstant: 25),
-            terraTitleLabel.widthAnchor.constraint(equalToConstant: 100)
-        ])
-    }
-    
-    private func setSubtitleLabelConstraints() {
-        NSLayoutConstraint.activate([
-            subtitleLabel.topAnchor.constraint(equalTo: terraTitleLabel.bottomAnchor, constant: 20),
-            subtitleLabel.leadingAnchor.constraint(equalTo: terraTitleLabel.leadingAnchor),
-            subtitleLabel.heightAnchor.constraint(equalToConstant: 20),
-            subtitleLabel.widthAnchor.constraint(equalToConstant: 300)
-        ])
-    }
-    
-    private func setCriticalSpeciesLabelConstraints() {
-        NSLayoutConstraint.activate([
-            criticalSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
-            criticalSpeciesLabel.topAnchor.constraint(equalTo: scrollView.topAnchor, constant: 20),
-            criticalSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
-            criticalSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
-            
-        ])
-    }
-    
-    private func setCriticalSpeciesCVConstraints() {
-        NSLayoutConstraint.activate([
-            criticalSpeciesCV.topAnchor.constraint(equalTo: criticalSpeciesLabel.bottomAnchor, constant: 30),
-            criticalSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
-            criticalSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
-            
-        ])
-    }
-    
-    private func setEndangeredSpeciesLabelConstraints() {
-        NSLayoutConstraint.activate([
-            endangeredSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
-            endangeredSpeciesLabel.topAnchor.constraint(equalTo: criticalSpeciesCV.bottomAnchor, constant: 30),
-            endangeredSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
-            endangeredSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
-        ])
-    }
-
-    private func setEndangeredSpeciesCVConstraints() {
-        NSLayoutConstraint.activate([
-            endangeredSpeciesCV.topAnchor.constraint(equalTo: endangeredSpeciesLabel.bottomAnchor, constant: 30),
-            endangeredSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
-            endangeredSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
-            endangeredSpeciesCV.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor)
-        ])
-    }
-
-    private func setVulnerableSpeciesLabelConstraints() {
-           NSLayoutConstraint.activate([
-               vulnerableSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
-               vulnerableSpeciesLabel.topAnchor.constraint(equalTo: endangeredSpeciesCV.bottomAnchor, constant: 30),
-               vulnerableSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
-               vulnerableSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
-           ])
-       }
-
-       private func setVulnerableSpeciesCVConstraints() {
-           NSLayoutConstraint.activate([
-               vulnerableSpeciesCV.topAnchor.constraint(equalTo: vulnerableSpeciesLabel.bottomAnchor, constant: 30),
-               vulnerableSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
-               vulnerableSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
-               vulnerableSpeciesCV.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor)
-           ])
-       }
 }
 
 //MARK: -- CollectionView Methods
@@ -263,7 +153,6 @@ extension SpeciesListViewController: UICollectionViewDataSource {
         case vulnerableSpeciesCV: return filteredVulnerableSpecies.count
         default: return 0
         }
-        
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
@@ -321,9 +210,122 @@ extension SpeciesListViewController: UICollectionViewDelegateFlowLayout {
     
 }
 
+//MARK: --ScrollView Methods
 extension UIScrollView {
     func updateContentView() {
-        
         contentSize.height = subviews.sorted(by: { $0.frame.maxY < $1.frame.maxY }).last?.frame.maxY ?? contentSize.height
+    }
+}
+
+
+//MARK: -- Adding subviews/constraints
+extension SpeciesListViewController {
+    
+    private func addSubviews() {
+        let parentViewUIElements = [terraTitleLabel, subtitleLabel, scrollView]
+        parentViewUIElements.forEach { view.addSubview($0) }
+        parentViewUIElements.forEach { $0.translatesAutoresizingMaskIntoConstraints = false }
+        
+        let scrollViewUIElements = [criticalSpeciesLabel, criticalSpeciesCV, endangeredSpeciesLabel, endangeredSpeciesCV, vulnerableSpeciesLabel, vulnerableSpeciesCV]
+        scrollViewUIElements.forEach{ scrollView.addSubview($0) }
+        scrollViewUIElements.forEach{ $0.translatesAutoresizingMaskIntoConstraints = false }
+    }
+    
+    private func setConstraints() {
+        setTerraTitleLabelConstraints()
+        setSubtitleLabelConstraints()
+        
+        setCriticalSpeciesLabelConstraints()
+        setCriticalSpeciesCVConstraints()
+        
+        setEndangeredSpeciesLabelConstraints()
+        setEndangeredSpeciesCVConstraints()
+        
+        setVulnerableSpeciesLabelConstraints()
+        setVulnerableSpeciesCVConstraints()
+    }
+    
+    private func setScrollViewConstraints(){
+        scrollView.backgroundColor = .clear
+        NSLayoutConstraint.activate([
+            scrollView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            scrollView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            scrollView.topAnchor.constraint(equalTo: subtitleLabel.bottomAnchor, constant: 30),
+            scrollView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+        ])
+    }
+    
+    
+    private func setTerraTitleLabelConstraints() {
+        NSLayoutConstraint.activate([
+            terraTitleLabel.topAnchor.constraint(equalTo: view.topAnchor, constant: 80),
+            terraTitleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 30),
+            terraTitleLabel.heightAnchor.constraint(equalToConstant: 25),
+            terraTitleLabel.widthAnchor.constraint(equalToConstant: 100)
+        ])
+    }
+    
+    private func setSubtitleLabelConstraints() {
+        NSLayoutConstraint.activate([
+            subtitleLabel.topAnchor.constraint(equalTo: terraTitleLabel.bottomAnchor, constant: 20),
+            subtitleLabel.leadingAnchor.constraint(equalTo: terraTitleLabel.leadingAnchor),
+            subtitleLabel.heightAnchor.constraint(equalToConstant: 20),
+            subtitleLabel.widthAnchor.constraint(equalToConstant: 300)
+        ])
+    }
+    
+    private func setCriticalSpeciesLabelConstraints() {
+        NSLayoutConstraint.activate([
+            criticalSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
+            criticalSpeciesLabel.topAnchor.constraint(equalTo: scrollView.topAnchor, constant: 20),
+            criticalSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
+            criticalSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
+            
+        ])
+    }
+    
+    private func setCriticalSpeciesCVConstraints() {
+        NSLayoutConstraint.activate([
+            criticalSpeciesCV.topAnchor.constraint(equalTo: criticalSpeciesLabel.bottomAnchor, constant: 30),
+            criticalSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
+            criticalSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
+            
+        ])
+    }
+    
+    private func setEndangeredSpeciesLabelConstraints() {
+        NSLayoutConstraint.activate([
+            endangeredSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
+            endangeredSpeciesLabel.topAnchor.constraint(equalTo: criticalSpeciesCV.bottomAnchor, constant: 30),
+            endangeredSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
+            endangeredSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
+        ])
+    }
+    
+    private func setEndangeredSpeciesCVConstraints() {
+        NSLayoutConstraint.activate([
+            endangeredSpeciesCV.topAnchor.constraint(equalTo: endangeredSpeciesLabel.bottomAnchor, constant: 30),
+            endangeredSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
+            endangeredSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
+            endangeredSpeciesCV.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor)
+        ])
+    }
+    
+    private func setVulnerableSpeciesLabelConstraints() {
+        NSLayoutConstraint.activate([
+            vulnerableSpeciesLabel.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor, constant: 30),
+            vulnerableSpeciesLabel.topAnchor.constraint(equalTo: endangeredSpeciesCV.bottomAnchor, constant: 30),
+            vulnerableSpeciesLabel.heightAnchor.constraint(equalToConstant: 30),
+            vulnerableSpeciesLabel.widthAnchor.constraint(equalToConstant: 300)
+        ])
+    }
+    
+    private func setVulnerableSpeciesCVConstraints() {
+        NSLayoutConstraint.activate([
+            vulnerableSpeciesCV.topAnchor.constraint(equalTo: vulnerableSpeciesLabel.bottomAnchor, constant: 30),
+            vulnerableSpeciesCV.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
+            vulnerableSpeciesCV.heightAnchor.constraint(equalToConstant: 235),
+            vulnerableSpeciesCV.centerXAnchor.constraint(equalTo: scrollView.centerXAnchor)
+        ])
     }
 }
